gcp/requirements.txt
functions-framework==3.*
keras==3.6.0
tensorflow-cpu==2.17.0
google-cloud-storage==2.18.2
Pillow==10.0.0
numpy==1.26.4
protobuf==4.25.3
